prediction_service/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=eigen
absl-py=0.11.0=pyhd3eb1b0_1
aiohttp=3.7.4=py38h2bbff1b_1
argon2-cffi=20.1.0=py38h2bbff1b_1
astor=0.8.1=pypi_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py38haa95532_0
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=py38haa95532_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2021.1.19=haa95532_0
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py38haa95532_0
cffi=1.14.5=py38hcd4344a_0
chardet=3.0.4=py38haa95532_1003
click=7.1.2=pyhd3eb1b0_0
cloudpickle=1.6.0=pypi_0
colorama=0.4.4=pyhd3eb1b0_0
coverage=5.4=py38h2bbff1b_2
cryptography=2.9.2=py38h7a1dbc1_0
cycler=0.10.0=pypi_0
cython=0.29.22=py38hd77b12b_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.6.0=pyhd3eb1b0_0
entrypoints=0.3=py38_0
frb=1.1.4=pypi_0
gast=0.4.0=py_0
google-auth=1.27.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
graphviz=2.38=hfd603c8_2
grpcio=1.35.0=py38hc60d5dd_1
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=2.0.0=py_1
importlib_metadata=2.0.0=1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.20.0=py38hd4e2768_1
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.17.2=py38haa95532_1
jinja2=2.11.3=pyhd3eb1b0_0
joblib=1.0.0=pypi_0
jpeg=9b=hb83a4c4_2
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.2.0=py_0
jupyter_core=4.7.1=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
keras=2.4.3=0
keras-applications=1.0.8=py_1
keras-base=2.4.3=py_0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=pypi_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.14.0=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
llvmlite=0.35.0=pypi_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.3.4=py38haa95532_0
markupsafe=1.1.1=py38he774522_0
matplotlib=3.3.3=pypi_0
mistune=0.8.4=py38he774522_1000
mkl=2020.2=256
mkl-service=2.3.0=py38h196d8e1_0
mkl_fft=1.3.0=py38h46781fe_0
mkl_random=1.1.1=py38h47e9c7a_0
msys2-conda-epoch=20160418=1
multidict=5.1.0=py38h2bbff1b_2
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.2=pyhd3eb1b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
notebook=6.2.0=py38haa95532_0
numba=0.52.0=pypi_0
numpy=1.19.2=py38hadc3359_0
numpy-base=1.19.2=py38ha3acd2a_0
oauthlib=3.1.0=py_0
openssl=1.1.1j=h2bbff1b_0
opt_einsum=3.1.0=py_0
packaging=20.9=pyhd3eb1b0_0
pandas=1.2.0=pypi_0
pandoc=2.11=h9490d1a_0
pandocfilters=1.4.3=py38haa95532_1
parso=0.7.0=py_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.0=pypi_0
pip=21.0.1=py38haa95532_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.14.0=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pydot=1.4.1=pypi_0
pygments=2.8.0=pyhd3eb1b0_0
pyjwt=2.0.1=py38haa95532_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38ha925a31_4
pyreadline=2.1=py38_1
pyrsistent=0.17.3=py38he774522_0
pysocks=1.7.1=py38haa95532_0
python=3.8.8=hdbf39b2_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python-dotenv=0.17.0=pypi_0
pytz=2020.5=pypi_0
pywin32=227=py38he774522_1
pywinpty=0.5.7=py38_0
pyyaml=5.4.1=py38h2bbff1b_1
pyzmq=20.0.0=py38hd77b12b_1
qt=5.9.7=vc14h73c81de_0
qtconsole=5.0.2=pyhd3eb1b0_0
qtpy=1.9.0=py_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.0=pypi_0
scipy=1.6.1=py38h14eb087_0
seaborn=0.11.1=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38haa95532_0
shap=0.38.1=pypi_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py38haa95532_0
sklearn=0.0=pypi_0
slicer=0.0.7=pypi_0
sqlite=3.33.0=h2a8f88b_0
ta-lib=0.4.19=pypi_0
tensorboard=2.3.0=pyh4dce500_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.3.0=mkl_py38h8c0d9a2_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-docs=0.0.05c49a035557c4cb3349d6b05ab5e6f5bd032fc32-=pypi_0
tensorflow-estimator=2.3.0=pyheb71bc4_0
termcolor=1.1.0=py38haa95532_1
terminado=0.9.2=py38haa95532_0
testpath=0.4.4=pyhd3eb1b0_0
threadpoolctl=2.1.0=pypi_0
tornado=6.1=py38h2bbff1b_0
tqdm=4.58.0=pypi_0
traitlets=5.0.5=pyhd3eb1b0_0
twitterapi=2.6.10=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.3=pyhd3eb1b0_0
vadersentiment=3.3.2=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
wrapt=1.12.1=py38he774522_1
yaml=0.2.5=he774522_0
yarl=1.6.3=py38h2bbff1b_0
zeromq=4.3.3=ha925a31_3
zipp=3.4.0=pyhd3eb1b0_0
zlib=1.2.11=h62dcd97_4
